+++ alerter/src/data_store/stores/alert.py
@@ -8,6 +8,8 @@ import pika.exceptions
 from src.data_store.mongo.mongo_api import MongoApi
 from src.data_store.redis.store_keys import Keys
 from src.data_store.stores.store import Store
+from src.alerter.alert_severities import Severity
+from src.alerter.alert_code import InternalAlertCode
 from src.message_broker.rabbitmq.rabbitmq_api import RabbitMQApi
 from src.utils.constants import (STORE_EXCHANGE, HEALTH_CHECK_EXCHANGE,
                                  ALERT_STORE_INPUT_QUEUE,
@@ -119,32 +121,109 @@ class AlertStore(Store):
         $max is the timestamp of the last alert entered
         $inc increments n_alerts by one each time an alert is added
         """
-        self.mongo.update_one(
-            alert['parent_id'],
-            {
-                'doc_type': 'alert',
-                'n_alerts': {'$lt': 1000}
-            }, {
-                '$push': {
-                    'alerts': {
-                        'origin': alert['origin_id'],
-                        'alert_name': alert['alert_code']['name'],
-                        'severity': alert['severity'],
-                        'message': alert['message'],
-                        'metric': alert['metric'],
-                        'timestamp': str(alert['timestamp']),
-                    }
-                },
-                '$min': {'first': alert['timestamp']},
-                '$max': {'last': alert['timestamp']},
-                '$inc': {'n_alerts': 1},
-            }
-        )
+
+        # Do not save the internal alerts into Mongo as they aren't useful to
+        # the user
+        if alert['severity'] != Severity.INTERNAL.value:
+            self.mongo.update_one(
+                alert['parent_id'],
+                {
+                    'doc_type': 'alert',
+                    'n_alerts': {'$lt': 1000}
+                }, {
+                    '$push': {
+                        'alerts': {
+                            'origin': alert['origin_id'],
+                            'alert_name': alert['alert_code']['name'],
+                            'severity': alert['severity'],
+                            'message': alert['message'],
+                            'metric': alert['metric'],
+                            'timestamp': str(alert['timestamp']),
+                        }
+                    },
+                    '$min': {'first': alert['timestamp']},
+                    '$max': {'last': alert['timestamp']},
+                    '$inc': {'n_alerts': 1},
+                }
+            )
 
     def _process_redis_store(self, alert: Dict) -> None:
-        metric_data = {'severity': alert['severity'],
-                       'message': alert['message']}
-        key = alert['origin_id']
-        self.redis.hset(Keys.get_hash_parent(alert['parent_id']),
-                        eval('Keys.get_alert_{}(key)'.format(alert['metric'])),
-                        json.dumps(metric_data))
+        if alert['severity'] == Severity.INTERNAL.value:
+            if (alert['alert_code']['code'] ==
+                InternalAlertCode.ComponentResetAll.value and
+                    alert['origin_id'] == 'SystemAlertersManager'):
+
+                """
+                The `ComponentResetAll` alert indicates that PANIC has
+                started, or restarted. This means that we cannot be sure if the
+                configurations are the same as the previous run and the Alert
+                Metrics should be cleared for each CHAIN.
+                """
+                parent_hash = Keys.get_hash_parent_raw()
+                chain_hashes_list = self.redis.get_keys_unsafe(
+                            '*' + parent_hash + '*')
+
+                # Go through all the chains that are in REDIS
+                for chain in chain_hashes_list:
+                    # For each chain we need to load all the keys and only
+                    # delete the ones that match the pattern `alert_system*`
+                    # REDIS doesn't support this natively
+                    for key in self.redis.hkeys(chain):
+                        # We only want to delete alert keys
+                        if 'alert_system' in key and self.redis.hexists(chain,
+                                                                        key):
+                            self.redis.hremove(chain, key)
+
+            elif (alert['alert_code']['code'] ==
+                  InternalAlertCode.ComponentReset.value and
+                  alert['origin_id'] == 'SystemAlertersManager'):
+
+                """
+                This internal alert is sent whenever an Alerter is
+                terminated due to change in configuration or shut down signal.
+                """
+                # For the specified chain we need to load all the keys and only
+                # delete the ones that match the pattern `alert_system*`
+                # REDIS doesn't support this natively
+                chain_hash = Keys.get_hash_parent(alert['parent_id'])
+                for key in self.redis.hkeys(chain_hash):
+                    # We only want to delete alert keys
+                    if 'alert_system' in key and self.redis.hexists(chain_hash,
+                                                                    key):
+                        self.redis.hremove(chain_hash, key)
+
+            elif (alert['alert_code']['code'] ==
+                  InternalAlertCode.ComponentReset.value and
+                  alert['origin_id'] == 'GithubAlerterManager'):
+                """
+                The `ComponentReset` alert for the `GithubAlerterManager`
+                indicates that PANIC has started, or restarted. This means
+                that we cannot be sure if the configurations are the same as
+                the previous run and the Alert Metrics should be cleared for
+                each CHAIN.
+                """
+                parent_hash = Keys.get_hash_parent_raw()
+                chain_hashes_list = self.redis.get_keys_unsafe(
+                            '*' + parent_hash + '*')
+
+                # Go through all the chains that are in REDIS
+                for chain in chain_hashes_list:
+                    # For each chain we need to load all the keys and only
+                    # delete the ones that match the pattern `alert_github2*`
+                    # REDIS doesn't support this natively
+                    for key in self.redis.hkeys(chain):
+                        # We only want to delete alert_github2 which
+                        # corresponds to the metric `cannot_access_github`
+                        if 'alert_github2' in key and self.redis.hexists(chain,
+                                                                         key):
+                            self.redis.hremove(chain, key)
+        else:
+            # If the alert is not of severity Internal, the metric needs to be
+            # stored in REDIS
+            metric_data = {'severity': alert['severity'],
+                           'message': alert['message']}
+            key = alert['origin_id']
+            self.redis.hset(
+                Keys.get_hash_parent(alert['parent_id']),
+                eval('Keys.get_alert_{}(key)'.format(alert['metric'])),
+                json.dumps(metric_data))