+++ docs/CHANGE_LOG.md
@@ -2,6 +2,16 @@ Change the contents of this file to this:
 
 # Change Log
 
+## 0.1.3 [DRAFT]
+
+Released on _
+
+- Fixed tests to work with alerter changes. Merged multiple tests into one using parameterization
+- Updated Alerter to fix bugs with regards to metric changes in thresholds
+- Added Internal Alerts on startup originating from the Alerter, which are used to reset all metrics for that chain
+- Added functionality to cater for new Internal Alert in Data Store
+- Added Tests for new Internal Alerts in System/Github Alerter and Alert Store
+
 ## 0.1.2
 
 Released on 25th March 2021